Fix default pager
Fixes #20:

It turns out with old pager `less -F` were exiting `less` when there are
less than a screenful of entries, and exit clears the screen by default
in less. So we either need to remove `-F` or add `-X`. Latter would
leave a lot of text on the screen (including the case where history
doesn't fit the screen), so this commit implements the former. Default
pager is now just `less -R` (`-F` removed)