Better prefix of error message
If you're using edgedb CLI in the scripts, ocasionally you may see
somethign like this:
```
Error: authentication failed
```
This is not very clear whether it comes from edgedb CLI or any other
tool. Even if you enable `set +x` (i.e. tracing command run) in shell it
may still be confusing to run `docker run edgedb-image edgedb` or
`sudo edgedb`.

So it's quite common for unix tools to put their command name before the
error message. This PR adds it consistently in fatal errors and few
other places but not in interactive usage, as in interactive case it
should be clean enough which utility is currently active.