Do not block Docker installation method on missing/inaccessible socket
There is no Docker deamon Unix socket on Windows (there's a TCP port and
a named pipe, though [1]).  But, more importantly, the CLI itself
doesn't use the socket in any way, so the check doesn't seem to be
useful.  Replace it with a CLI check in the form of calling
`docker version`.

[1] https://docs.docker.com/desktop/faqs/#how-do-i-connect-to-the-remote-docker-engine-api