repl: Add print for used configuration
This also introduces a way for color printing:
1. Styles are semantic, no colors thorough code
2. Don't need intermediate strings (`x.to_string().light_blue()`)
3. Disabling colors works even when formatting to a string
4. Themes can be used later

Also introduces new print macro `echo` which:
1. Doesn't require format pattern (patterns have little use with colors)
2. Buffers the whole line before printing (friendlier to other
   processes, faster)

This unfortunately introduces another way for color printing thorought
code, but we eventually have to convert colors to semantic names for
theme support anyway.

Other than thate, this color priting is mostly in proof of concept state
as more styles should be added to make it useful.