Change fallback socket dir (#583)
Previously, on linux if there is no XDG_RUNTIME_DIR, we checked wether
`/run/user/<userid>` exists, and used that as a default for
XDG_RUNTIME_DIR, otherwise data dir is used.

This creates an issue for the command-line tool, as now it has to check
two locations for the sockets. But unlike for server-side client must
poll both locations during `wait-until-available` period, because
technically directory can appear at any time.

But generally it makes little sense, since absense of XDG_RUNTIME_DIR is
almost always means there is no `systemd --user` daemon (unless
environment was cleared out).

Instead of guessing runtime directory this commit uses the same fallback
as we use on MacOS: always use `~/.cache/edgedb/run/<name>` as a socket
directory if XDG_RUNTIME_DIR is not set (this also means there is no
fallback to putting sockets in the data dir, since we assume cache dir
is always possible to create).