macos: bootout inactive services too
Previously it was assumed that if service is not running it's not loaded
in launchctl, which is obviously not true.

Currently we assume service is not loaded when `launchctl print` returns
error. Which seems to be correct, but may have exceptions. Still the
hypothesis is that it's better then checking stderr text.

Fixes #845