cloud login: Obtain long-lived secret key explicitly
The CLI auth cycle returns a short-lived key which should be used to
obtain a proper long-lived secret key if necessary, so do just that.

Also, avoid re-logging-in if a valid secret key is already present.  We
might need to consider adding `--force` to initiate a new login (or just
recommend doing `edgedb cloud logout` to force login).