Reduce code duplication, handle history mismatch cases more granularly
Move common database migration history code to the new `db_migration`
module to reduce duplication and make it more like the filesystem
history (although some more unification is due).

Handle more cases of database/fs history mismatch, such as when the
database history is ahead of filesystem, if histories diverge, or if
they are completely unrelated.