Add info for current binary and install paths
This allows getting both the containing path to the currently executing
binary under `binary-dir`, but also the path that the self-installer
would install the binary to under `install-dir`.

There are some slight naming confusions happening here since we were
calling the install path the "binary path". Open to doing a naming
refactor here to clean that up.