Fix for initial installation when the CLI does the arch switch on macOS
When the CLI installs itself on macOS it checks if the correct binary
architecture was downloaded (i.e. aarch64 for Apple Silicon), and if the
binary is x86, it calls the `upgrade` path instead to download the
correct binary.  Unfortunately, the `upgrade` path assumes the
installation directory already exists, and it doesn't on the initial
install, so the whole thing fails with "No such file or directory".

Fix this by making the directories on upgrade too.