Fix force_database_error and get_current_branch
When opening REPL, we query `sys::get_current_database`.
If `force_database_error` is set, this query fail, preventing even
entering the REPL, which is bad for obvious reasons.

I've made sure to override `force_database_error` is these queries
and even made it so if we already have branch set in connection
options, we don't even make the query, but just use that query param
as the current branch.

Closes #1321