Remove one style of printing from CLI (#1393)
echo! was a macro that had its own print style which made it difficult to understand. A few uses of echo! incorrectly assumed it took format!-style args, which was actually incorrect.

Most of this PR was done in a mechanical way via this grit script, and then touched up by hand, and then a pass where msg! was replaced by println! temporarily was used to give clippy a chance to inline format args.