Install new executable name (and link old executable) (#1398)
This installs (when --features gel is passed) to both edgedb and gel executables. The alternative executable is hard linked, symlinked or copied (in that order of preference).

Note that the gel executable is installed even w/o --features gel right now, but it is installed as the "alternate" executable.

We also check the name of the executable that is invoked and warn if the alternative is there. If the two executables are out of sync (determined by checking executable length), we warn as well.