Unify --instance arg for extension commands
Extension commands used to have an "instance" paramater
that used same keys (--instance, -I) as the global instance connection
parameter. This PR unifies these two.

The practical effect is that `-I`/`--instance` can be specified anywhere
between commands instead of just at the end, as it used to have to be.

We might also want to infer the instance from the current project.