Implement a tabular output format and make it the default for SQL (#1460)
It uses the Printer/Formatter mechanism, since that is what actually
has fully implemented formatters.

I would like to take advantage of the coloring that Printer/Formatter
does, but the table library we use breaks if there are control
characters. I have a plan I might do to fork the library and get
something working, but that would be a follow-up.

This adds a new tabular output-format that works for both edgeql
and SQL. We then add a sql-output-format mode that defaults to
tabular, but it can still be set back to default.

Fixes #1451.