Use edgedb/edgeql mirror instead of the main repo
The CLI depends on the main github.com/edgedb/edgedb repo for the
tokenizer of EdegQL in the edgeql-parser crate.

When cargo resolves a git dependency it does a recursive clone.
There is currenctly no way around this.

So the build of the cli has to clone the main repo and all git submodules,
which includes `postgres`, `libpg_query` and a few smaller repos. None of
them are actaully needed.

---

To get around that, I've setup a mirror repo, with a script that clones
the main repo and extract only the edgeql-parser crate
(and one derive subcrate).

This mirror repo must be updated manually for now, but this is not a big deal,
as changes to edgeql-parser are sparse currently.