Fix extension package command default channel and slot (#1482)
Default channel should probably be Stable, since that is where we
publish postgis (even when built against an rc). Default slots need to
be formatted like `6-rc-2`, *not* `6-rc2`, since that is how it is in
the `server_slot` field. (I think this is a mismatch with what "slots"
we use for the server builds themselves?)

I will probably try to do another follow-up that does more flexible
matching against slots.